+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- [DesignToolbox] Display versions using OUDS tag in about menu (Orange-OpenSource/ouds-ios#1203)
 - [DesignToolbox] Define timeout of 2 hours for all GitHub Actions workflows
 - [DesignToolbox] Display of colors semantic tokens (Orange-OpenSource/ouds-ios#1206)
 - [DesignToolbox] Display of colors and size semantic tokens (Orange-OpenSource/ouds-ios#1206)